Updated requirements.txt
There were two dependencies (bleach and psutil) that github's
dependabot flagged as having security problems so I'm updating the
requirements.txt file to get the newer versions in.

I am using a new-ish repository that only has the dependencies for one
post, though, so I'm not updating the requirements.in file that is
more complete (it might have dependencies that I don't want anymore so
I'm also not installing everything in it).